requirements.txt
fastapi
uvicorn[standard]
weaviate-client
openai
pydantic
python-dotenv
requests
streamlit
aiohttp
pyyaml
pytest
pytest-asyncio
pytest-mock
pytest-cov
httpx
